refactor(filters): adapt all filters to package-level implementations
