+++ filters/horizontal_flip/pipeline.go
@@ -1,0 +1,9 @@
+package horizontal_flip
+
+import "dougdomingos.com/image-filters/filters/types"
+
+var HorizontalFlipPipeline = types.FilterPipeline{
+	Preprocess:       nil,
+	SerialFilter:     serialHorizontalFlip,
+	ConcurrentFilter: concurrentHorizontalFlip,
+}
