+++ filters/vertical_flip/pipeline.go
@@ -1,0 +1,9 @@
+package vertical_flip
+
+import "dougdomingos.com/image-filters/filters/types"
+
+var VerticalFlipPipeline = types.FilterPipeline{
+	Preprocess: nil,
+	SerialFilter: serialVerticalFlip,
+	ConcurrentFilter: concurrentVerticalFlip,
+}
